+++ src/main.js
@@ -1,4 +1,6 @@
 import { createRouter } from "./router/router";
+import { setupEventListeners } from "./managers/eventManager.js";
 
 const root = document.getElementById("root");
-createRouter(root);
+const { navigate } = createRouter(root);
+setupEventListeners(navigate);
