+++ src/router/router.js
@@ -14,6 +14,7 @@ export const createRouter = (root) => {
   const handleRoute = () => {
     const path = window.location.pathname;
 
+    //라우트 가드
     if (path === "/profile" && !localStorage.getItem("user")) {
       navigate("/login");
       return;
@@ -29,53 +30,11 @@ export const createRouter = (root) => {
   };
 
   const navigate = (path) => {
+    // 브라우저 url 업데이트
     window.history.pushState({}, "", path);
+    // 페이지 렌더링
     handleRoute();
   };
-
-  // 전역 이벤트 리스너 등록
-  document.addEventListener("click", (e) => {
-    if (e.target.id === "logout") {
-      e.preventDefault();
-      localStorage.removeItem("user");
-      navigate("/login");
-      return;
-    }
-
-    if (e.target.matches("[data-link]")) {
-      e.preventDefault();
-      navigate(e.target.getAttribute("href"));
-    }
-  });
-
-  document.addEventListener("submit", (e) => {
-    if (e.target.id === "profile-form") {
-      e.preventDefault();
-      const formData = new FormData(e.target);
-      const user = JSON.parse(localStorage.getItem("user") || "{}");
-      const updatedUser = {
-        username: formData.get("username") || user.username,
-        email: formData.get("email") || user.email,
-        bio: formData.get("bio") || user.bio,
-      };
-      localStorage.setItem("user", JSON.stringify(updatedUser));
-      navigate("/profile");
-    }
-
-    if (e.target.id === "login-form") {
-      e.preventDefault();
-      const username = e.target.querySelector("#username").value;
-      localStorage.setItem(
-        "user",
-        JSON.stringify({
-          username,
-          email: "",
-          bio: "",
-        }),
-      );
-      navigate("/");
-    }
-  });
 
   window.addEventListener("popstate", handleRoute);
   handleRoute();
